requirements.txt
torch==2.9.1
transformers==4.57.1
peft==0.18.0
datasets==4.4.1
pyyaml==6.0.3
safetensors==0.6.2
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.7
seaborn==0.13.2
huggingface-hub==0.36.0
